Remove unecessary function wrappers in plot data code
