+++ packages/studio-base/src/panels/Plot/plotData.ts
@@ -7,9 +7,14 @@ import memoizeWeak from "memoize-weak";
 
 import { Time } from "@foxglove/rostime";
 import { Immutable as Im } from "@foxglove/studio";
+import {
+  MessageAndData,
+  MessageDataItemsByPath,
+} from "@foxglove/studio-base/components/MessagePathSyntax/useCachedGetMessagePathDataItems";
 import { getDatasetsFromMessagePlotPath } from "@foxglove/studio-base/panels/Plot/datasets";
 import { Bounds, makeInvertedBounds, unionBounds } from "@foxglove/studio-base/types/Bounds";
 import { Range } from "@foxglove/studio-base/util/ranges";
+import { getTimestampForMessage } from "@foxglove/studio-base/util/time";
 
 import {
   BasePlotPath,
@@ -132,6 +137,27 @@ function compare(a: Im<PlotData>, b: Im<PlotData>): number {
   const startCompare = rangeA.start - rangeB.start;
   return startCompare !== 0 ? startCompare : rangeA.end - rangeB.end;
 }
+
+function mapPathItem(messageAndData: MessageAndData) {
+  const headerStamp = getTimestampForMessage(messageAndData.messageEvent.message);
+  return {
+    queriedData: messageAndData.queriedData,
+    receiveTime: messageAndData.messageEvent.receiveTime,
+    headerStamp,
+  };
+}
+
+/**
+ * Fetch the data we need from each item in itemsByPath and discard the rest of
+ * the message to save memory.
+ */
+export const getByPath = (itemsByPath: MessageDataItemsByPath): PlotDataByPath => {
+  const ret: PlotDataByPath = {};
+  for (const [path, items] of Object.entries(itemsByPath)) {
+    ret[path] = items.map(mapPathItem);
+  }
+  return ret;
+};
 
 /**
  * Reduce multiple PlotData objects into a single PlotData object, concatenating messages
